refactor: relocate AgentAvatar to components/Agents
Move the AgentAvatar display component (and its private LazyAgentAvatar) out of utils/agents.tsx into client/src/components/Agents/AgentAvatar.tsx. utils/ is reserved for pure helpers; a stateful component with hooks did not belong there.

The remaining utilities (isImageCached, getAgentAvatarUrl, getContactDisplayName) are JSX-free, so utils/agents.tsx is renamed to utils/agents.ts. The component test cases move to components/Agents/tests/AgentAvatar.spec.tsx; pure-function tests move to utils/__tests__/agents.spec.ts.

Updates four import sites (AgentCard, AgentDetail, AgentDetailContent, FavoriteItem) and two test mocks (Accessibility, AgentDetail) to point at the new module path.
