+++ client/src/components/Agents/AgentCard.tsx
@@ -2,8 +2,9 @@ import React, { useMemo, useState } from 'react';
 import { Label, OGDialog, OGDialogTrigger } from '@librechat/client';
 import type t from 'librechat-data-provider';
 import { useLocalize, TranslationKeys, useAgentCategories } from '~/hooks';
-import { cn, AgentAvatar, getContactDisplayName } from '~/utils';
+import { cn, getContactDisplayName } from '~/utils';
 import AgentDetailContent from './AgentDetailContent';
+import AgentAvatar from './AgentAvatar';
 
 interface AgentCardProps {
   agent: t.Agent;
